Added ability to send/accept invites to Pools
Added ability for Pool Managers to send invites to Pools and for Users
to accept or reject them.  Created UI for editing Pools and wired up
controllers to them.  Adjusted security rules to add support for the
invitation system.

Added firebase-util library.  This was a quick workaround for the time
being but needs to be replaced with a filter-based solution.

closes #3 and #16